+++ crates/medea-reactive/src/collections/hash_map.rs
@@ -12,7 +12,7 @@ use futures::stream::{LocalBoxStream, StreamExt as _};
 
 use crate::subscribers_store::{
     common, progressable,
-    progressable::{RecheckableCounterFuture, RecheckableFutureExt},
+    progressable::{AllProcessed, Processed},
     SubscribersStore,
 };
 
@@ -123,8 +123,7 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_insert_processed(&self) -> RecheckableCounterFuture {
+    pub fn when_insert_processed(&self) -> Processed<'static, ()> {
         self.on_insert_subs.when_all_processed()
     }
 
@@ -133,8 +132,7 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_remove_processed(&self) -> RecheckableCounterFuture {
+    pub fn when_remove_processed(&self) -> Processed<'static, ()> {
         self.on_remove_subs.when_all_processed()
     }
 
@@ -143,11 +141,10 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_all_processed(&self) -> impl RecheckableFutureExt<Output = ()> {
-        crate::join_all(vec![
-            self.when_remove_processed(),
-            self.when_insert_processed(),
+    pub fn when_all_processed(&self) -> AllProcessed<'static, ()> {
+        crate::when_all_processed(vec![
+            self.when_remove_processed().into(),
+            self.when_insert_processed().into(),
         ])
     }
 }
@@ -222,7 +219,7 @@ where
         ))
     }
 
-    /// Chains [`HashMap::replay_on_insert`] with a [`HashMap::on_insert`].
+    /// Chains [`HashMap::replay_on_insert()`] with a [`HashMap::on_insert()`].
     #[inline]
     pub fn on_insert_with_replay(&self) -> LocalBoxStream<'static, O> {
         Box::pin(self.replay_on_insert().chain(self.on_insert()))