+++ crates/medea-reactive/src/collections/hash_set.rs
@@ -6,7 +6,7 @@ use futures::stream::LocalBoxStream;
 
 use crate::subscribers_store::{
     common, progressable,
-    progressable::{RecheckableCounterFuture, RecheckableFutureExt},
+    progressable::{AllProcessed, Processed},
     SubscribersStore,
 };
 
@@ -121,8 +121,7 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_insert_processed(&self) -> RecheckableCounterFuture {
+    pub fn when_insert_processed(&self) -> Processed<'static, ()> {
         self.on_insert_subs.when_all_processed()
     }
 
@@ -131,8 +130,7 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_remove_processed(&self) -> RecheckableCounterFuture {
+    pub fn when_remove_processed(&self) -> Processed<'static, ()> {
         self.on_remove_subs.when_all_processed()
     }
 
@@ -141,11 +139,10 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_all_processed(&self) -> impl RecheckableFutureExt<Output = ()> {
-        crate::join_all(vec![
-            self.when_remove_processed(),
-            self.when_insert_processed(),
+    pub fn when_all_processed(&self) -> AllProcessed<'static, ()> {
+        crate::when_all_processed(vec![
+            self.when_remove_processed().into(),
+            self.when_insert_processed().into(),
         ])
     }
 }