+++ crates/medea-reactive/src/collections/vec.rs
@@ -6,7 +6,7 @@ use futures::stream::LocalBoxStream;
 
 use crate::subscribers_store::{
     common, progressable,
-    progressable::{RecheckableCounterFuture, RecheckableFutureExt},
+    progressable::{processed::AllProcessed, Processed},
     SubscribersStore,
 };
 
@@ -104,8 +104,7 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_push_processed(&self) -> RecheckableCounterFuture {
+    pub fn when_push_processed(&self) -> Processed<'static, ()> {
         self.on_push_subs.when_all_processed()
     }
 
@@ -114,8 +113,7 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_remove_processed(&self) -> RecheckableCounterFuture {
+    pub fn when_remove_processed(&self) -> Processed<'static, ()> {
         self.on_remove_subs.when_all_processed()
     }
 
@@ -124,11 +122,10 @@ where
     ///
     /// [`Future`]: std::future::Future
     #[inline]
-    #[must_use]
-    pub fn when_all_processed(&self) -> impl RecheckableFutureExt<Output = ()> {
-        crate::join_all(vec![
-            self.when_remove_processed(),
-            self.when_push_processed(),
+    pub fn when_all_processed(&self) -> AllProcessed<'static, ()> {
+        crate::when_all_processed(vec![
+            self.when_remove_processed().into(),
+            self.when_push_processed().into(),
         ])
     }
 }