+++ crates/medea-reactive/src/field/mod.rs
@@ -18,7 +18,7 @@ use futures::{
 };
 
 use crate::subscribers_store::{
-    progressable, progressable::RecheckableCounterFuture, SubscribersStore,
+    progressable, progressable::Processed, SubscribersStore,
 };
 
 #[doc(inline)]
@@ -134,7 +134,7 @@ where
     /// subscribers.
     ///
     /// [`Future`]: std::future::Future
-    pub fn when_all_processed(&self) -> RecheckableCounterFuture {
+    pub fn when_all_processed(&self) -> Processed<'static, ()> {
         self.subs.when_all_processed()
     }
 }