+++ crates/medea-reactive/CHANGELOG.md
@@ -23,12 +23,13 @@ All user visible changes to this project will be documented in this file. This p
     - `ProgressableVec`;
     - `ProgressableHashMap`;
     - `ProgressableHashSet`.
-- `ProgressableCell` container ([#164]).
+- `ProgressableCell` container ([#170]);
+- `Processed` and `AllProcessed` futures ([#170]).
 
 [#81]: /../../pull/81
 [#103]: /../../pull/103
 [#159]: /../../pull/159
-[#164]: /../../pull/164
+[#170]: /../../pull/170